+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Release Notes
 
+## Additions
+
+- Added a teamless flag in the features
+
 ## Improvements
 
-- Added trimming to all empty string checks
-- Updated packages
+- Prevented repeated onboarding
+- Renamed enabledFeatures to features
+- Updated Packages
+
+## Fixes
+
+- Fixed teamless team name issue